CRDCDH-539 Migrate to independent component This will help reduce merge conflicts
