+++ src/graphql/index.ts
@@ -48,6 +48,9 @@ export type { Response as ListBatchesResp } from "./listBatches";
 export { query as LIST_LOGS } from "./listLogs";
 export type { Response as ListLogsResp } from "./listLogs";
 
+export { mutation as VALIDATE_SUBMISSION } from "./validateSubmission";
+export type { Response as ValidateSubmissionResp } from "./validateSubmission";
+
 // User Profile
 export { query as GET_USER } from "./getUser";
 export type { Response as GetUserResp } from "./getUser";
